Rank2BinaryForms

Documentation for Rank2BinaryForms. Package to compute the roots of the polynomials in the image of the map \Gamma defined in ArXiv 1901.08320.

Rank2BinaryForms.Image_affineGamma

Rank2BinaryForms.affineGamma

Rank2BinaryForms.buildpolynomial

Rank2BinaryForms.crossratio

Rank2BinaryForms.set_getting_rootsofunity

Rank2BinaryForms.triplets

Rank2BinaryForms.tuples

Rank2BinaryForms.Image_affineGamma — Method

Image_affineGamma(d)

Returns the roots of all the polynomials in the image of Γ.

Rank2BinaryForms.affineGamma — Method

affineGamma(d)

Returns a function that computes the unknown roots of the polynomial Γ(triplet).

Rank2BinaryForms.buildpolynomial — Method

buildpolynomial(roots, vars)

Generic function to build the polynomials in the image of Γ from the list of roots.

Rank2BinaryForms.crossratio — Method

crossratio(A, B, C, D)

Returns the cross ratio of four numbers. There is also the method:

crossratio(v) = crossratio(v...)

Rank2BinaryForms.set_getting_rootsofunity — Method

set_getting_rootsofunity(f)

Sets the method used to compute the d-th roots of unity. The input is a function f computing the roots of unity accepting calls as f(d)(k) and returning the k-th root of the d-th roots of unity. The order of the d-th roots of unity can be clockwise or anti-clockwise, but they must be ordered. Examples of valid inputs are the following functions f and g:

    g(d) = k -> exp(2*pi*im*k/d)
    using Nemo
    CC = ComplexField(64);
    f(d) = k -> Nemo.root_of_unity(CC, d)^k;

Rank2BinaryForms.triplets — Method

triplets(d)

Returns a vector of triplets [i,j,k] of integers 1 <= i < j < k <= d, up to the action of the d-th dihedral group.

Rank2BinaryForms.tuples — Method

tuples(d)

Returns a vector of vectors resulting to pushfirst each integer 1:d in triplet.