Rank2BinaryForms

Documentation for Rank2BinaryForms. Package to compute the roots of the polynomials in the image of the map \Gamma defined in ArXiv 1901.08320.

Rank2BinaryForms._buildpolynomial

Rank2BinaryForms.bc_canonical_to_any

Rank2BinaryForms.crossratio

Rank2BinaryForms.fix_homo_base!

Rank2BinaryForms.four_tuples

Rank2BinaryForms.gamma

Rank2BinaryForms.gamma_image

Rank2BinaryForms.get_ratios

Rank2BinaryForms.homo_uroots

Rank2BinaryForms.minors_indices

Rank2BinaryForms.set_getting_rootsofunity

Rank2BinaryForms.standard_triplets

Rank2BinaryForms.uroots

Rank2BinaryForms._buildpolynomial — Method

_buildpolynomial(roots, vars, _)

Generic function to build the polynomials in the image of Γ from the list of roots.

Rank2BinaryForms.bc_canonical_to_any — Function

bc_canonical_to_any(L)
bc_canonical_to_any(L, R)

Returns PGL matrix sending e_i=[0,..,1,..,0] to v[i] and [1,...,1] to u.

n = 100
v1 = matrix(QQ, rand(-n:n, 4, 1))
v2 = matrix(QQ, rand(-n:n, 4, 1))

u = 2*v1-4*v2
normalize_mat(u, v1, v2) == reduce(hcat, (2.*v1, -4.*v2))
true

Rank2BinaryForms.crossratio — Method

crossratio(A, B, C, D)

Returns the cross ratio of four numbers.

Rank2BinaryForms.fix_homo_base! — Method

fix_homo_base!(M, u)

Returns

Rank2BinaryForms.four_tuples — Method

four_tuples()

Returns a function on the standard triplets of integers I, which returns tuples [[i, I...] for i in 1:d but not in I.

Rank2BinaryForms.gamma — Method

gamma()

Returns a function on triplets of integers I which computes the roots of the polynomial Γ(I).

Rank2BinaryForms.gamma_image — Method

gamma_image()

Returns the roots of all the polynomials in the image of Γ.

Rank2BinaryForms.get_ratios — Method

get_ratios(u, M)

Given a vector u being a linear combination of the columns of M, returns the coefficients of such a linear combination. That is, returns v=[a_1,..,a_n] such that Mv = u.

``julia n = 100 v1 = matrix(QQ, rand(-n:n, 4, 1)) v2 = matrix(QQ, rand(-n:n, 4, 1))

u = 2v1-4v2 get_ratios([v1,v2], u) [ 2] [-4] ```

Rank2BinaryForms.homo_uroots — Function

homo_uroots()
homo_uroots(I)

Returns vectors [1,-ξ_i] for the i-th (i in I = 1:d) d-th-root of unity.

Rank2BinaryForms.minors_indices — Method

minors_indices(M, k)

Returns the indices (as an iterator) for all the kxk minors of M.

Rank2BinaryForms.set_getting_rootsofunity — Method

set_getting_rootsofunity(f)

Sets the method used to compute the d-th roots of unity. The input is a function f computing the roots of unity accepting calls as f(d)(k) and returning the k-th root of the d-th roots of unity. The order of the d-th roots of unity can be clockwise or anti-clockwise, but they must be ordered. Examples of valid inputs are the following functions f and g:

    g(d, k) = exp(2*pi*im*k/d)
    set_getting_rootsofunity(g)

    using Nemo
    CC = ComplexField(64);
    f(d, k) = Nemo.root_of_unity(CC, d)^k;
    set_getting_rootsofunity(f)

Rank2BinaryForms.standard_triplets — Method

standard_triplets()

Returns a vector with all the standard triplets of integers [1,j,k] with 1 < j < k <= d.

Rank2BinaryForms.uroots — Function

uroots()
uroots(I)

Returns the vector [ξ_i...] where ξ_i is the i-th d-th-root of unity for i in I (default I = 1:d).